perf: cut peak memory in mmap-cache save, lassosum2 LD, and inference r2
Three confirmed memory wins from an efficiency audit (no behaviour change):

- mmap LD cache save: streamed each block straight into the memmap sidecar via
  np.lib.format.open_memmap instead of np.concatenate(parts) + astype, which held
  ~2x the payload in RAM while writing. Peak extra RAM is now one block, so a
  many-GB LD cache no longer needs a second full copy to be written.
- lassosum2: keep the LD blocks in float32 (was float64) -- the LD rows are the
  memory/bandwidth-dominant part and float32 correlations are plenty precise;
  beta/Rb/accumulators stay float64, so each product still promotes to float64.
  Halves lassosum2's resident LD.
- inference r2: keep the cross-chain R*b products (the Rb dict, the memory hog of
  the r2 step: n_chains x n_saved x m) in float32 instead of casting samples to
  float64. The r2 is a Monte-Carlo correlation estimate whose sampling noise
  dwarfs float32; halves that buffer (~1.6 GB -> 0.8 GB at 1M SNPs / 10 chains).

202 tests pass (mmap roundtrip, lassosum2, inference covered).

+++ ldpred3/infer.py
@@ -294,7 +294,10 @@ def ldpred3_auto_infer(corr, beta_hat, n_eff, *, n_chains=10,
         # apply_R already uses the shrunk LD (_prep_corr / _prep_blocks applied
         # shrink_corr once); do NOT shrink again here, or the dense path would
         # use shrink^2*R + (1-shrink^2)*I and disagree with the block path.
-        prod = apply_R(s.astype(float))            # (n_saved, m) = R_shrunk b
+        # Keep the products in float32 (the samples' dtype) — this Rb dict is the
+        # memory hog of the r² step (n_chains × n_saved × m), and the r² is a
+        # Monte-Carlo correlation estimate whose sampling noise dwarfs float32.
+        prod = apply_R(s)                          # (n_saved, m) float32 = R_shrunk b
         Rb[c] = prod
     r2_vals = []
     kept_with_samples = [c for c in keep if c in Rb]
